refactor: Login 엔티티를 Auth 엔티티로 변경 및 FSD 구조 변경
