+++ src/views/list-plan/index.tsx
@@ -6,7 +6,7 @@ import Spacer from '@/src/shared/ui/Spacer'
 import BlankTooltip from '@/src/features/plan/blank-tooltip'
 import { PlanType, useGetPlans } from '@/src/entities/plan'
 import CardPlan from '@/src/features/plan/card-plan'
-import { getLoginUrl } from '@/src/entities/login/api'
+import { getLoginUrl } from '@/src/entities/auth'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
 import { useGetMyProfile } from '@/src/entities/user'
